radix_tree: optimize emplace
Change bottom_leaf to return any leaf placed lower than a certain depth.
This way, leaf variable in internal_emplace will always point to an
element with specified key (if only it exists). It saves going descending
the tree for the second time when the element is an embedded entry.